movie rating is average of every reviews back

+++ src/main/java/fr/but3/cinehub/entity/Movie.java
@@ -22,7 +22,7 @@ public class Movie {
     private Long id;
     private String name;
     private Integer duration;
-    private int rating;
+    private double rating;
     private String directors;
     private String actors;
     private String synopsis;
